+++ app/services/evidence_check_selector.rb
@@ -19,7 +19,6 @@ class EvidenceCheckSelector
   end
 
   def evidence_check_type
-    return 'random'
     if random_evidence_check?
       'random'
     elsif flagged?
